bug fixes and cleanup of multi-threaded alpha-beta
also a cleanup of declarations and casts

+++ client/boardspace-codename1/boardspace core/bridge/URLEncoder.java
@@ -194,7 +194,7 @@ public class URLEncoder {
             throw new NullPointerException("charsetName");
 
         for (int i = 0; i < s.length();) {
-            int c = (int) s.charAt(i);
+            int c = s.charAt(i);
             //System.out.println("Examining character: " + c);
             if (dontNeedEncoding.get(c)) {
                 if (c == ' ') {
@@ -222,7 +222,7 @@ public class URLEncoder {
                           + " is high surrogate");
                         */
                         if ( (i+1) < s.length()) {
-                            int d = (int) s.charAt(i+1);
+                            int d = s.charAt(i+1);
                             /*
                               System.out.println("\tExamining "
                               + Integer.toHexString(d));
@@ -239,7 +239,7 @@ public class URLEncoder {
                         }
                     }
                     i++;
-                } while (i < s.length() && !dontNeedEncoding.get((c = (int) s.charAt(i))));
+                } while (i < s.length() && !dontNeedEncoding.get((c = s.charAt(i))));
 
                 charArrayWriter.flush();
                 String str = new String(charArrayWriter.toCharArray());
