+++ client/boardspace-codename1/boardspace core/online/game/commonCanvas.java
@@ -7843,7 +7843,7 @@ public abstract class commonCanvas extends exCanvas
     		int mod = Math.min(w,h)/2;
     		if(mod>0)
     			{
-    			int step = (int)(flashState%mod);
+    			int step = (flashState%mod);
     			flashState++;
     			GC.frameRect(gc,((step&1)==0)?Color.blue:Color.red,left-step,top-step,w+step*2,h+step*2);
     			}
