+++ client/boardspace-codename1/boardspace games/bugs/BugsViewer.java
@@ -518,8 +518,8 @@ public class BugsViewer extends CCanvas<BugsCell,BugsBoard> implements BugsConst
     	CELLSIZE = (int)cs;
     	//G.print("cell "+cs0+" "+cs+" "+bestPercent);
     	// center the board in the remaining space
-    	int boardW = (int)(ncols*CELLSIZE);
-    	int boardH = (int)(nrows*CELLSIZE);
+    	int boardW = (ncols*CELLSIZE);
+    	int boardH = (nrows*CELLSIZE);
     	
 
     	int extraW = Math.max(0,(mainW-boardW)/2);
