+++ client/boardspace-codename1/boardspace games/circle/CircleViewer.java
@@ -325,8 +325,8 @@ public class CircleViewer extends CCanvas<CircleCell,CircleBoard> implements Cir
     	CELLSIZE = (int)cs;
     	//G.print("cell "+cs0+" "+cs+" "+bestPercent);
     	// center the board in the remaining space
-    	int boardW = (int)(ncols*CELLSIZE);
-    	int boardH = (int)(nrows*CELLSIZE);
+    	int boardW = (ncols*CELLSIZE);
+    	int boardH = (nrows*CELLSIZE);
     	int extraW = Math.max(0, (mainW-boardW)/2);
     	int extraH = Math.max(0, (mainH-stateH*2-boardH)/2);
     	int boardX = mainX+extraW;
