+++ client/boardspace-codename1/boardspace games/crossfire/CrossfireViewer.java
@@ -265,8 +265,8 @@ public class CrossfireViewer extends CCanvas<CrossfireCell,CrossfireBoard> imple
     	CELLSIZE = (int)cs;
     	//G.print("cell "+cs0+" "+cs+" "+bestPercent);
     	// center the board in the remaining space
-    	int boardW = (int)(nrows*CELLSIZE);
-    	int boardH = (int)(nrows*CELLSIZE);
+    	int boardW = (nrows*CELLSIZE);
+    	int boardH = (nrows*CELLSIZE);
     	int extraW = Math.max(0, (mainW-boardW-CELLSIZE*3)/2);
     	int extraH = Math.max(0, (mainH-boardH)/2);
     	int boardX = mainX+extraW;
