+++ client/boardspace-codename1/boardspace games/euphoria/EuphoriaViewer.java
@@ -615,7 +615,7 @@ private Color playerBackground[] = {
     	int mainW = G.Width(main);
     	int mainH = G.Height(main);
     	// calculate a suitable cell size for the board
-    	double cs = Math.min((double)mainW/ncols,(double)(mainH/(nrows+1)));
+    	double cs = Math.min(mainW/ncols,(mainH/(nrows+1)));
     	bb.CELLSIZE = STANDARD_CELLSIZE = CELLSIZE = (int)cs;
     	// G.print("cell "+cs0+" "+cs+" "+bestPercent);
     	// center the board in the remaining space
@@ -1192,7 +1192,7 @@ private Color playerBackground[] = {
     	drawStackOnPlayer(gc,pl,gb,
     			gb.legalToHitPlayer(p.authority,sources,dests)?highlight:null,
     			pr,p.authority,
-    			xp-unitSize/3,G.Top(r)+(int)(unitSize*4),tip,fromHiddenWindow);
+    			xp-unitSize/3,G.Top(r)+(unitSize*4),tip,fromHiddenWindow);
     	boolean hit =false;
 		boolean hide = allPlayersLocal() && !fromHiddenWindow && (view==PlayerView.AutoArtifacts);
     	switch(view)
