+++ client/boardspace-codename1/boardspace games/magnet/MagnetViewer.java
@@ -276,8 +276,8 @@ public class MagnetViewer extends CCanvas<MagnetCell,MagnetBoard> implements Mag
 		CELLSIZE = (int)cs;
 		//G.print("cell "+cs0+" "+cs+" "+bestPercent);
 		// center the board in the remaining space
-		int boardW = (int)(ncols*CELLSIZE);
-		int boardH = (int)(nrows*CELLSIZE);
+		int boardW = (ncols*CELLSIZE);
+		int boardH = (nrows*CELLSIZE);
 		int extraW = Math.max(0, (mainW-boardW)/2);
 		int extraH = Math.max(0, (mainH-boardH-stateH*2)/2);
 		int boardX = mainX+extraW;
