+++ client/boardspace-codename1/boardspace games/mijnlieff/MijnlieffViewer.java
@@ -310,8 +310,8 @@ public class MijnlieffViewer extends CCanvas<MijnlieffCell,MijnlieffBoard> imple
     	CELLSIZE = (int)cs;
     	//G.print("cell "+cs0+" "+cs+" "+bestPercent);
     	// center the board in the remaining space
-    	int boardW = (int)(ncols*CELLSIZE);
-    	int boardH = (int)(nrows*(CELLSIZE-2));
+    	int boardW = (ncols*CELLSIZE);
+    	int boardH = (nrows*(CELLSIZE-2));
     	int extraW = Math.max(0, (mainW-boardW)/2);
     	int extraH = Math.max(0, (mainH-boardH)/2);
     	int boardX = mainX+extraW;
