+++ client/boardspace-codename1/boardspace games/qe/QEViewer.java
@@ -345,8 +345,8 @@ public boolean PerformAndTransmit(commonMove m, boolean transmit,replayMode repl
        	int C2 = CELLSIZE/2;
             	// center the board in the remaining space
         int stateH = fh*5/2;
-    	int boardW = (int)(ncols*CELLSIZE);
-    	int boardH = (int)((nrows-1)*CELLSIZE);
+    	int boardW = (ncols*CELLSIZE);
+    	int boardH = ((nrows-1)*CELLSIZE);
     	int extraW = (mainW-boardW)/2;
     	int extraH = (mainH-boardH)/2;
     	int boardX = mainX+extraW;
@@ -364,7 +364,7 @@ public boolean PerformAndTransmit(commonMove m, boolean transmit,replayMode repl
         {
             int l = boardX;
             int t = boardY;
-            int sz = CELLSIZE * (int)nrows;
+            int sz = CELLSIZE * nrows;
             G.SetRect(calculatorRect,CELLSIZE*3,CELLSIZE*3,CELLSIZE*7,CELLSIZE*10);
             G.SetRect(openBid,(int)(l+0.19*sz),(int)(t+0.58*sz),(int)(0.2*sz),(int)(0.2*sz));
             G.SetRect(reserveCards,(int)(l+0.19*sz),(int)(t+0.18*sz),(int)(0.2*sz),(int)(0.38*sz));
