+++ client/boardspace-codename1/boardspace games/takojudo/TakojudoViewer.java
@@ -194,8 +194,8 @@ public class TakojudoViewer extends CCanvas<TakojudoCell,TakojudoBoard> implemen
     	SQUARESIZE = (int)cs;
      	//G.print("cell "+cs0+" "+cs+" "+bestPercent);
     	// center the board in the remaining space
-    	int boardW = (int)(ncols*SQUARESIZE);
-    	int boardH = (int)(nrows*SQUARESIZE);
+    	int boardW = (ncols*SQUARESIZE);
+    	int boardH = (nrows*SQUARESIZE);
     	int extraW = Math.max(0, (mainW-boardW)/2);
     	int extraH = Math.max(0, (mainH-stateH*2-boardH)/2);
     	int boardX = mainX+extraW;
